backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
paddleocr>=2.7.3
paddlepaddle>=2.6.0
pillow>=10.0.0
numpy>=1.24.0